Fix parameter sqrt issue

+++ distkeras/workers.py
@@ -176,7 +176,7 @@ class MassWorker(NetworkWorker):
         # Set the worker id.
         self.set_worker_id(worker_id)
         # Prepare the gradient residual matrix.
-        v = np.asarray(self.model.get_weights(), dtype=np.float32)
+        v = np.asarray(self.model.get_weights())
         v.fill(0.0)
         # Start the epoch training process
         try:
@@ -197,9 +197,9 @@ class MassWorker(NetworkWorker):
                     self.pull()
                     # Update the local replica.
                     self.model.set_weights(self.center_variable)
-                W1 = np.asarray(self.model.get_weights(), dtype=np.float32)
+                W1 = np.asarray(self.model.get_weights())
                 self.model.train_on_batch(X, Y)
-                W2 = np.asarray(self.model.get_weights(), dtype=np.float32)
+                W2 = np.asarray(self.model.get_weights())
                 delta = W2 - W1
                 v += delta
                 self.iteration += 1
